@@ -45,22 +45,23 @@
     <purplepip.argLine>-Dlogback.configurationFile=${purplepip.log.file} -Dlogback.debug=${logback.debug} ${purplepip.argLine.extra}</purplepip.argLine>
 
     <!-- Java version -->
-    <java.version>1.8</java.version>
+    <java.version>11</java.version>
 
     <!-- Library versions -->
-    <antlr.version>4.7.1</antlr.version>
-    <assertj.version>3.8.0</assertj.version>
-    <guava.version>23.6-jre</guava.version>
+    <antlr.version>4.8-1</antlr.version>
+    <assertj.version>3.16.1</assertj.version>
+    <guava.version>29.0-jre</guava.version>
     <hamcrest.version>2.0.0.0</hamcrest.version>
-    <jacoco.version>0.8.0</jacoco.version>
+    <jacoco.version>0.8.5</jacoco.version>
     <jacoco.covered.ratio>0.75</jacoco.covered.ratio>
-    <junit.version>4.12</junit.version>
-    <kotlin.version>1.2.31</kotlin.version>
+    <junit.version>5.7.0-M1</junit.version>
+    <kotlin.version>1.3.72</kotlin.version>
     <logback.version>1.2.3</logback.version>
-    <lombok.version>1.16.20</lombok.version>
-    <metrics.version>4.0.0</metrics.version>
-    <mockito.version>2.13.0</mockito.version>
+    <lombok.version>1.18.12</lombok.version>
+    <metrics.version>4.1.10.1</metrics.version>
+    <mockito.version>3.3.3</mockito.version>
     <slf4j.version>1.7.25</slf4j.version>
+    <surefire.version>3.0.0-M5</surefire.version>
   </properties>
 
   <dependencyManagement>
